ocpp-v1_6: Change transactionId type from int to long

+++ ocpp-v1_6-test/src/main/java/eu/chargetime/ocpp/test/FakeCentralSystem.java
@@ -293,7 +293,7 @@ public class FakeCentralSystem {
     return result;
   }
 
-  public void sendRemoteStopTransactionRequest(int transactionId) throws Exception {
+  public void sendRemoteStopTransactionRequest(long transactionId) throws Exception {
     RemoteStopTransactionRequest request = new RemoteStopTransactionRequest();
     request.setTransactionId(transactionId);
     send(request);
